Massive improvements
Re-Wrote code that determines best hand from a set of cards. It is
now slightly less general but a lot faster.

Hand calculations now work correctly for low straights with the
addition of LowAce as a type of card that is included when looking
for straights when an Ace is present